Update scale factor and image size in training script
The scale factor for training images was lowered from 0.5 to 0.35, potentially improving the variety of scales on which the model trains. Also, the image size was increased from 192x192 to 256x256, which might bring better training results due to larger image details. These changes may contribute to improved model performance.