Split main method from train.py into new file train_celeba.py
The main method used for training the model was moved from train.py into a new file called train_celeba.py. This change is designed to enhance the organization and readability of the code. This will also simplify potential changes or updates to the training process in the future by isolating the main training procedure into a separate file.