Add pixel_only parameter and MergedCCPDataset class
The 'pixel_only' parameter has been added to the initialization of datasets. This dictates whether the path list is merged with the mask path list and the label path list. Furthermore, a new class named 'MergedCCPDataset' has been introduced that includes additional functionality for masks. It facilitates the merging of different category masks into a combined mask.