Change 'has_pixel_labels' from boolean to integer
Modified the 'has_pixel_labels' variable in datasets.py to store integer type values instead of boolean. This change was made in the process of loading and processing the image labels and also during the determination of whether each image index has pixel labels or not. This is expected to improve the performance.