Increase batch size and adjust counter condition
The batch size in DataLoader within the train_deepfashion2 file has been increased from 8 to 16 for train_dataset. In train.py, the condition which checks for counter value has been updated to >-1 from >=-1 ensuring counter takes only positive value and accuracy logging is in sync with counter value, not epoch.