Refactor image processing in dataset and train scripts
Updated the image processing in 'datasets.py' to accommodate a reverse input for image size. This change mostly involved an adjustment in the initialization of the image size and crop dimensions. Modifications were also made to the mask generation process, as well as the image cropping and resizing logic. Data splitting in 'train_deepfashion2.py' was commented out. Correspondingly, the invocation of the image displaying function in 'train.py' was also commented.