Implement training function for DeepFashion2 and modify processing in dataset
Added a function to train the DeepFashion2 model in 'train.py'. This function includes logic for training segmentation, classification or the whole network based on the mode. In 'datasets.py', image size initialization and crop dimensions were updated, along with modifications to mask generation and image cropping/resizing. In 'train_deepfashion2.py', data splitting was commented out. The invocation of the image displaying function in 'train.py' was also commented out.