Add IntegratedModel for multi-task learning
The model utilizes a shared ResNet50 backbone for object detection, image segmentation, and image classification. To accomplish this, a RetinaNet Head is included for detection, a U-Net like decoder is used for segmentation, and a simple average pooling and a linear layer is utilized for classification. Also, necessary functions for calculation of performance metrics like mAP, F1 score and IOU are added. The code for debugging images was commented out and an unused variable was removed.