Update training scripts and add augmented dataset
The update includes modifications made to the train.py script such as the removal of the input mode check in the 'train_DeepFashion2' function and changes to how mAP, f1, and IOU values are averaged. A new function 'val_DeepFashion2' was also added for validation purposes. Alongside this, AugmentedDeepFashion2Dataset class was added in datasets.py to generate augmented datasets. Modifications were made to the 'train_deepfashion2.py' script to utilize the augmented dataset in training operations.