Update resnet model and simplify train function call
The resnet model has been updated from resnet101 to resnet152 in models.py to possibly improve the prediction accuracy, and also the train function call has been simplified to be more readable in train_deepfashion2.py. This simplification does not alter any functionality but improves the readability and understandability of the code.