Refactor code and optimise image segmentation
The image segmentation function has been reduced in size and moved to another file to streamline the code and improve its organisation. Furthermore, the number of data loads for the training dataset has been increased, which is expected to enhance the model's performance. Used Resnet50 instead of Resnet101 to improve speed.