Refactor F1 score calculation in training and validation
Changed the F1 score calculation in the training and validation process to use binary conversion before calculation. This improves the F1 metric accuracy as it works more effectively with binary labels for multi-label classification problems. Also, num_workers in DataLoader is set to 0 for both training and validation process in 'train_deepfashion2.py' to remove multi-processing.