Adjust loss calculation and optimizer learning rate
In this commit, the final bounding box loss value in the training script is scaled by a factor of 4. Additionally, commented out the calculation conversion of attributes and prediction classes to numpy arrays. In the DeepFashion2 training script, the learning rate of the Adam optimizer has been reduced from 5e-4 to 1e-4.