Update model structure and refine metrics calculation function
The commit updates the model to use resnet50 instead of resnet18, refactors the Decoder structure to match this change, and adds an extension to the bounding box generation part of the model. Additionally, the 'calculate_map' function has been improved to focus on boxes associated with label 1, and its use has been adapted across the training scripts. The commit also changes the model saving directory.