Adjust true boxes selection and validate accuracy calculation
The selection of relevant true boxes in train.py now considers only the indices where true labels == 1, which improves accuracy of training. In train_deepfashion2.py, the validation accuracy is now correctly calculated as the average of average precision (mAP) and F1 score, and the image size is adjusted to 226x300.