Update bounding box and loss calculations in various scripts
This commit pertains to the improvement of bounding box adjustments and updates in the 'datasets.py' script. Moreover, the calculations for losses in the 'train.py' script have been optimized, outdating redundant lines. In addition, enhancements have been made to the display of category labels in the 'utils.py' script for better visualization, and adjustments have been made to the 'test_deepfashion2.py' script including the change of the model directory path and correct handling of prediction classes.